BUG: df.replace with numeric values and string to_replace
